refactor: rename code-generators to generators and update package structure
