+++ app/build.gradle.kts
@@ -2,7 +2,7 @@ dependencies {
     // Internal
     implementation(project(":dsl"))
     implementation(project(":common"))
-    implementation(project(":code-generators"))
+    implementation(project(":generators"))
     // External
 
 }
