--- code-generators/README.md
@@ -1,3 +1,0 @@
-# PIEKER Code Generators Module
-
-TODO
