+++ generators/src/main/java/pieker/generators/code/PiekerCodeGenerationException.java
@@ -1,4 +1,4 @@
-package pieker.code.generators;
+package pieker.generators.code;
 
 public class PiekerCodeGenerationException extends RuntimeException {
     public PiekerCodeGenerationException(String message) {
