+++ generators/src/main/java/pieker/generators/code/VelocityTemplateProcessor.java
@@ -1,4 +1,4 @@
-package pieker.code.generators;
+package pieker.generators.code;
 
 import org.apache.velocity.Template;
 import org.apache.velocity.VelocityContext;
@@ -22,7 +22,7 @@ public class VelocityTemplateProcessor {
 
     public VelocityTemplateProcessor() {
         this.templateDirectory = "templates/";
-        this.outputDirectory = ".gen/";
+        this.outputDirectory = "../.gen/";
         this.velocityEngine = initVelocityEngine();
     }
