+++ generators/src/main/java/pieker/generators/code/multistep/package-info.java
@@ -2,4 +2,4 @@
  * This package contains the classes that are used to generate the code for multistep test components, like multistep
  * proxies which the PIEKER supervisor can trigger to change their behavior.
  */
-package pieker.code.generators.multistep;
+package pieker.generators.code.multistep;
