+++ generators/src/main/java/pieker/generators/components/docker/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the classes that are responsible for generating Docker components for the PIEKER system, such
+ * as Dockerfiles and Images for generated test components.
+ */
+package pieker.generators.components.docker;
