+++ generators/src/main/resources/templates/traffic/thread.vm
