+++ settings.gradle.kts
@@ -2,6 +2,6 @@ rootProject.name = "pieker"
 
 include("app")
 include("architectures")
-include("code-generators")
+include("generators")
 include("common")
 include("dsl")
